[MISC] Refactor vectorised invocation of banded and unbanded alignment algorithm.
The banded and unbanded function call operators were basically the same.
The respective parts that are in common are factored out into separate pre and postprocess functions.
This had a little effect on the alignment microbenchmarks also no structural changes happened.
The coordinate computation has been adapted to return the correct value for each alignment.