[MISC] Move aligned sequence builder construction into main algorithm.
The algorithm determines how the memory of the matrices is used during the alignment computation.
Accordingly, only the algorithm can determine where the global optimum points to in the respective trace matrix.
In the banded alignment only a fraction of the original matrix is allocated and matrix index of the cell representing the optimum
needs to be mapped to the actual trace matrix index.
The alignment builder policy would also need to get this information.
To make the interface more clean, the aligned sequence builder construction is
put into the scope of the alignment algorithm, because the alignment algorithm can access all internally stored policy data.
However, the alignment algorithm only returns a lambda function,
which, when invoked by the alignment result builder,
generates the aligned sequence builder and calls it with the
respective data.