[MISC] Merge the trace iterator implementation.
The trace iterator was split to differentiate between iterating through
a banded and an unbanded trace matrix.
In the new banded alignment computation the same
iterator could basically be used to iterate through the
banded and the unbanded matrix.
Also differentiating between the types always resulted in
extra overhead to differentiate in the calling code.
This commit cleans up this split by merging the implementations and offering only one type.