WIP Improve vectorised runtime for global alignment
This is now a more optimised version of the alignment:

It combines the best score and the horizonal column in one pair such they are loaded together into the cache
The simd iota view is now a random acces iterator.
In the column computation the column is tracked separately from the computation.
In the global alignment only the last cell is tracked.
This allows to spare the iteration over the index matrix.