[MISC] Refactor the selection of the alignment matrix policy.
Adds a dedicated transformation trait to configure the alignment matrix policy.
Makes the coordinate matrix a template parameter of the alignment class.
Removes unused template parameters such as the matrix index type and the score type.
Later the score type should be replaced by a member type exposed
by the alignment matrix.
The alignment matrix could be first combined inside
of the policy as well.