IMPROVEMENT: Improve access to uncompressed ibf
Adds aligned memory manager to the sdsl int vector allocation. This is only WIP and needs to be more refined.
In general the SDSL needs a way to specify the memory manager, or to allow aligned memory allocation.

The bulk_contains now reads an entire memory region as big as the simd vector from the uncompressed ibf and directly stores it into the binnin_vector.
Thus, no extra gather/scatter is needed to load the ibf, compute the bitwise-AND, and store the results again in the binning bitvector.
This is hard coded and only works with AVX512 at the moment.

It gives roughly another 1.5x speedup over the gather/scatter version before.