[NOP] Move the counting vector into its own header file.
The counting vector is somewhat independent of the interleaved bloom filter and should be placed into its own header file.
This keeps the header file for the interleaved bloom filter more smaller and easier to maintain.

The counting vector will be specialised such that it can also run in simd mode with AVX512.