[FEATURE] Enable AVX512 based couting using mask_add operation.
This adds a new function to get the simd counting agent.
This function is only availble if the corresponding AVX512 instruction set exists.
The user does not see the changes except of a different type.
The interfaces are seprated to give the user some control
over which parts should be using simd registers.
The counting vector implements now also a operator+= for bit vectors using the _mm512_mask_add_epi instruction for the different word sizes.